pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
    "requests>=2.32.5",
    "telegram>=0.0.1",
    "web3>=7.13.0",
]
